\name{AAFT}
\alias{AAFT}
\title{AAFT}
\description{use fourier phase shuffling on Xcorrelogram and
rank ordering to retain the distribution of Xdistribution
but here transform to gaussian, shuffle phases and retransform}
\usage{AAFT(X, method = c("shell"))}
\arguments{
  \item{X}{
}
  \item{method}{
}
}



\author{Fabian Gans, Miguel Mahecha, Max-Planck-Institute for Biogeochemistry, Jena}





