\name{buildPkg}
\alias{buildPkg}
\title{buildPkg}
\description{create the zip file of the package, that can be installed from R-GUI.}
\usage{buildPkg(pkg = getPackageId())}
\arguments{
  \item{pkg}{
}
}
\details{Run from windows it will be a zip file 
while run from a nix system it will be a tgz file relating to the architecture.}


\author{Thomas Wutzler}





