\name{copy2clip}
\alias{copy2clip}
\title{copy2clip}
\description{copies argument to the clipboard}
\usage{copy2clip(x, col.names = NA, row.names = FALSE, quote = FALSE, 
    sep = "\\t", ...)}
\arguments{
  \item{x}{
}
  \item{col.names}{
}
  \item{row.names}{
}
  \item{quote}{
}
  \item{sep}{
}
  \item{\dots}{
}
}

\value{x again}

\author{Thomas Wutzler}



\seealso{\link{twDev}}

