\name{fSampleFunc}
\alias{fSampleFunc}
\title{fSampleFunc}
\description{Adding two vectors}
\usage{
fSampleFunc(a.n.v,
b.n.v)
}
\arguments{
  \item{a.n.v}{numeric vector: first term}
  \item{b.n.v}{numeric vector: second term}
}
\details{If the two vectors are of different length, the shorter one
is recycled.}
\value{The componentwise sum of the two arguments}

\author{Thomas Wutzler}



\seealso{\code{\link{myPackageId}}}
\examples{
# This example appears in the generated help.
fSampleFunc(1:10, 2)
}
