\name{getPackageId}
\alias{getPackageId}
\title{getPackageId}
\description{Read the package id from the DESCRIPTION file in the current working directory.}
\usage{getPackageId(...)}
\arguments{
  \item{\dots}{further arguments to \code{\link{getPackageDesc}}}
}

\value{Scalar String: the package identifier}

\author{Thomas Wutzler}



\seealso{\link{twDev}}

