\name{iAAFT}
\alias{iAAFT}
\title{iAAFT}
\description{use fourier phase shuffling on Xcorrelogram and
rank ordering to retain the distribution of Xdistribution }
\usage{iAAFT(Xcor, Xdist = Xcor, tolerance = 0.01, maxit = 100, adjust.var = TRUE, 
    zero.mean = TRUE, quiet = TRUE, diff = FALSE, criterion = c("periodogram"), 
    rel.convergence = TRUE, method = c("shell"))}
\arguments{
  \item{Xcor}{
}
  \item{Xdist}{
}
  \item{tolerance}{
}
  \item{maxit}{
}
  \item{adjust.var}{
}
  \item{zero.mean}{
}
  \item{quiet}{
}
  \item{diff}{
}
  \item{criterion}{
}
  \item{rel.convergence}{
}
  \item{method}{
}
}
\details{This function iteratively generates a surrogate time series with the same spectrum and distribution }


\author{Henning Rust
Fabian Gans, Miguel Mahecha, Max-Planck-Institute for Biogeochemistry, Jena}





