\name{iAAFTSurrogateEnsemble}
\alias{iAAFTSurrogateEnsemble}
\title{Get random ensemble time series}
\description{calls the iAAFT function written by Henning Rust to get an ensemble of surrogate time series}
\usage{iAAFTSurrogateEnsemble(x, nsamples = 10)}
\arguments{
  \item{x}{input time series}
  \item{nsamples}{number of surrogate time series}
}

\value{Array containing a ensemble of time series with same distribution an spectrum as the input}

\author{Fabian Gans, MPI BGC Jena, Germany, fgans@bgc-jena.mpg.de
Fabian Gans, Miguel Mahecha, Max-Planck-Institute for Biogeochemistry, Jena}





