\name{installPkg}
\alias{installPkg}
\title{installPkg}
\description{running R CMD check}
\usage{installPkg(options = "--html", pkg = getPackageId())}
\arguments{
  \item{options}{further options}
  \item{pkg}{package name, default parse DESCRIPTION}
}



\author{Thomas Wutzler}




\examples{
if( FALSE ){
	installPkg()
}
}
