\name{loadDll}
\alias{loadDll}
\title{loadDll}
\description{Load DLL (Shared object) from src directory}
\usage{loadDll(dynFilenameLocal = file.path("src", paste(pkg, .Platform$dynlib.ext, 
    sep = "")), pkg = getPackageId())}
\arguments{
  \item{dynFilenameLocal}{
}
  \item{pkg}{the name of the package, may be provided for performance reasons.}
}



\author{Thomas Wutzler}





