\name{plotensembles}
\alias{plotensembles}
\title{plot the decomposed series}
\description{Function that creates plots of the surrogates of the low-freuquency component of the respiration and temprature data}
\usage{plotensembles(result)}
\arguments{
  \item{result}{list: Results list generated by getQ10}
}



\author{Fabian Gans, Miguel Mahecha, Max-Planck-Institute for Biogeochemistry, Jena}





