\name{reloadPkg}
\alias{reloadPkg}
\title{reloadPkg}
\description{Reloading a package}
\usage{reloadPkg(pkg, force = TRUE, ...)}
\arguments{
  \item{pkg}{name of the package (scalar character string)}
  \item{force}{logical: should a package be detached even though other attached packages depend on it?}
  \item{\dots}{other arguments to \code{\link{detach}}}
}



\author{Thomas Wutzler}





