\name{runRCheck}
\alias{runRCheck}
\title{runRCheck}
\description{running R CMD check}
\usage{runRCheck(options = "--no-manual", pkg = getPackageId())}
\arguments{
  \item{options}{further options}
  \item{pkg}{package name, default parse DESCRIPTION}
}



\author{Thomas Wutzler}




\examples{
if( FALSE ){
	runRCheck()
}
}
