\name{twWin}
\alias{twWin}
\title{twWin}
\description{Creates a new window with default size and par.}
\usage{twWin(width = 4.6, height = 3.2, pointsize = 10, record = TRUE, 
    xpos = 40, ypos = -80, ...)}
\arguments{
  \item{width}{
}
  \item{height}{
}
  \item{pointsize}{
}
  \item{record}{
}
  \item{xpos}{
}
  \item{ypos}{
}
  \item{\dots}{
}
}

\value{the device that was active before opening the new window}

\author{Thomas Wutzler}



\seealso{\link{twDev}, \link{dev.set}}

