\name{RSCAPE-package}
\alias{RSCAPE-package}
\alias{RSCAPE}
\docType{package}
\title{R SCAPE package}
\description{Package for applying the SCAPE method to estimate soil Q10 values as described in ...}
\details{
\tabular{ll}{Package: \tab RSCAPE\cr
Title: \tab R SCAPE package\cr
Version: \tab 0.0.0\cr
Date: \tab 2013-05-06\cr
Author: \tab Fabian Gans\cr
Maintainer: \tab Fabian Gans <fgans@bgc-jena.mpg.de>\cr
Depends: \tab R (>= 2.14.0), spectral.methods, EMD, plyr, zoo, Rssa, forecast\cr
Suggests: \tab \cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr}
}
\author{Fabian Gans}

\keyword{ package }


